+++ src/memory_measure.rs
@@ -1,10 +1,10 @@
 pub struct MemoryMeasure {
-    pub time_seconds: u64,
+    pub time_seconds: f64,
     pub quantity: u64
 }
 
 impl MemoryMeasure {
-    pub fn new(s: u64, q: u64) -> Self {
+    pub fn new(s: f64, q: u64) -> Self {
         MemoryMeasure {time_seconds: s, quantity: q}
     }
 }
